requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
attrs==23.2.0
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
discord.py==2.3.2
Flask==3.0.3
frozenlist==1.4.1
gunicorn==22.0.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
multidict==6.0.5
packaging==24.1
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.2.1
Werkzeug==3.0.3
yarl==1.9.4